backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
PyYAML==6.0.1
uvicorn[standard]==0.24.0

# Database
pymongo==4.6.0
motor==3.3.2

# Authentication - Clerk
python-multipart==0.0.6
PyJWT==2.10.1
cryptography==41.0.7

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Providers
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pillow==10.1.0
python-magic==0.4.27

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging
structlog==23.2.0